requirements.txt
asgiref==3.7.2
attrs==23.2.0
boto3==1.34.89
botocore==1.34.89
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
debugpy==1.8.0
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-storages==1.14.2
django_advance_thumbnail==1.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
isodate==0.6.1
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
mdurl==0.1.2
mysqlclient==2.2.4
packaging==24.0
pillow==10.3.0
pip-autoremove==0.10.0
pip-tools==7.4.1
protobuf==5.26.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
